game-chat: refactor zklogin and app API auth integration
