+++ apps/game-chat/src/App.tsx
@@ -1,5 +1,6 @@
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom'
 import { Toaster } from 'sonner'
+import { AuthenticatedLayout } from './auth/AuthenticatedLayout'
 import { Authenticated } from './components/Authenticated'
 import { DuelLayout } from './duel/Duel'
 import { Landing } from './landing/Landing'
@@ -13,12 +14,14 @@ function App() {
     <BrowserRouter>
       <div className="w-screen h-screen">
         <Routes>
-          <Route path="/d/:slug" element={<Authenticated component={DuelLayout} />} />
-          <Route path="/d" element={<Authenticated component={WaitRoom} />} />
+          <Route index element={<Landing />} />
+          <Route element={<AuthenticatedLayout />}>
+            <Route path="/d/:slug" element={<Authenticated component={DuelLayout} />} />
+            <Route path="/d" element={<Authenticated component={WaitRoom} />} />
+            <Route path="/practice" element={<PracticeDuel />} />
+            <Route path="/welcome-reward" element={<Authenticated component={ClaimWelcomeReward} />} />
+          </Route>
           <Route path="/signin" element={<Signing />} />
-          <Route path="/practice" element={<PracticeDuel />} />
-          <Route path="/welcome-reward" element={<Authenticated component={ClaimWelcomeReward} />} />
-          <Route path="/" element={<Landing />} />
           <Route path="*" element={<Navigate to="/" />} />
         </Routes>
         <Toaster />
